requirements.txt
numpy
pillow
tensorflow
flask
openai==0.28
python-dotenv
tensorflow-cpu
easyocr==1.7.0
python-Levenshtein==0.21.0
python-dotenv
PyMuPDF
fuzzywuzzy